fix: description focus 시에 앱에 메세지 보내도록 수정
